# openAPI/central_sequence_service_1_0_0.yaml
openapi: 3.0.3
info:
  title: Central Sequence Service
  description: >-
    This API manages the assignment and updating of sequence numbers for various
    elements within a story, ensuring logical order and consistency. Data is
    persisted in an SQLite database and synchronized with the central FountainAI
    Typesense Service. Collection creation is mandatory and verified at startup.
  version: 1.0.0
paths:
...
  /service-discovery:
    get:
      tags:
        - Service Discovery
      summary: Discover peer services
      description: >-
        Queries the API Gateway's lookup endpoint to resolve the URL of a
        specified service.
      operationId: getServiceDiscovery
      parameters:
        - description: Name of the service to discover
          required: true
          schema:
            title: Service Name
            type: string
            description: Name of the service to discover
          name: service_name
          in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
...
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
...